Replace getCenterOfShape
